[#53]refactor: Constant 로직 분리
