+++ app/analysis/parser/viz_node_generator/expr_node/constant.py
@@ -1,0 +1,10 @@
+import ast
+
+
+class Constant:
+
+    def __init__(self, node: ast.Constant):
+        self.node = node
+
+    def get_value(self):
+        return self.node.value
